+++ Content.Client/Corvax/TTS/TTSSystem.cs
@@ -22,8 +22,10 @@ public sealed class TTSSystem : EntitySystem
     [Dependency] private readonly AudioSystem _audio = default!;
 
     private ISawmill _sawmill = default!;
-    private readonly MemoryContentRoot _contentRoot = new();
+    private static MemoryContentRoot _contentRoot = new();
     private static readonly ResPath Prefix = ResPath.Root / "TTS";
+
+    private static bool _contentRootAdded;
 
     /// <summary>
     /// Reducing the volume of the TTS when whispering. Will be converted to logarithm.
@@ -40,8 +42,13 @@ public sealed class TTSSystem : EntitySystem
 
     public override void Initialize()
     {
+        if (!_contentRootAdded)
+        {
+            _contentRootAdded = true;
+            _res.AddRoot(Prefix, _contentRoot);
+        }
+
         _sawmill = Logger.GetSawmill("tts");
-        _res.AddRoot(Prefix, _contentRoot);
         _cfg.OnValueChanged(CCCVars.TTSVolume, OnTtsVolumeChanged, true);
         SubscribeNetworkEvent<PlayTTSEvent>(OnPlayTTS);
     }
@@ -50,7 +57,6 @@ public sealed class TTSSystem : EntitySystem
     {
         base.Shutdown();
         _cfg.UnsubValueChanged(CCCVars.TTSVolume, OnTtsVolumeChanged);
-        _contentRoot.Dispose();
     }
 
     public void RequestPreviewTTS(string voiceId)
@@ -77,14 +83,16 @@ public sealed class TTSSystem : EntitySystem
             .WithVolume(AdjustVolume(ev.IsWhisper))
             .WithMaxDistance(AdjustDistance(ev.IsWhisper));
 
+        var soundSpecifier = new ResolvedPathSpecifier(Prefix / filePath);
+
         if (ev.SourceUid != null)
         {
             var sourceUid = GetEntity(ev.SourceUid.Value);
-            // _audio.PlayEntity(audioResource.AudioStream, sourceUid, audioParams);
+            _audio.PlayEntity(audioResource.AudioStream, sourceUid, soundSpecifier, audioParams);
         }
         else
         {
-            // _audio.PlayGlobal(audioResource.AudioStream, audioParams);
+            _audio.PlayGlobal(audioResource.AudioStream, soundSpecifier, audioParams);
         }
 
         _contentRoot.RemoveFile(filePath);
@@ -92,7 +100,7 @@ public sealed class TTSSystem : EntitySystem
 
     private float AdjustVolume(bool isWhisper)
     {
-        var volume = MinimalVolume + SharedAudioSystem.GainToVolume(_volume);
+        var volume = MinimalVolume + SharedAudioSystem.GainToVolume(_volume*5);
 
         if (isWhisper)
         {
