+++ Content.Server/Corvax/TTS/VoiceMaskSystem.TTS.cs
@@ -1,5 +1,5 @@
-﻿using Content.Server.Corvax.TTS;
-using Content.Shared.Corvax.TTS;
+﻿using Content.Shared.Corvax.TTS;
+using Content.Shared.Inventory;
 using Content.Shared.VoiceMask;
 
 namespace Content.Server.VoiceMask;
@@ -8,13 +8,13 @@ public partial class VoiceMaskSystem
 {
     private void InitializeTTS()
     {
-        SubscribeLocalEvent<VoiceMaskComponent, TransformSpeakerVoiceEvent>(OnSpeakerVoiceTransform);
+        SubscribeLocalEvent<VoiceMaskComponent, InventoryRelayedEvent<TransformSpeakerVoiceEvent>>(OnSpeakerVoiceTransform);
         SubscribeLocalEvent<VoiceMaskComponent, VoiceMaskChangeVoiceMessage>(OnChangeVoice);
     }
 
-    private void OnSpeakerVoiceTransform(EntityUid uid, VoiceMaskComponent component, TransformSpeakerVoiceEvent args)
+    private void OnSpeakerVoiceTransform(EntityUid uid, VoiceMaskComponent component, InventoryRelayedEvent<TransformSpeakerVoiceEvent> args)
     {
-        args.VoiceId = component.VoiceId;
+        args.Args.VoiceId = component.VoiceId;
     }
 
     private void OnChangeVoice(Entity<VoiceMaskComponent> entity, ref VoiceMaskChangeVoiceMessage msg)
