+++ Content.Shared/Corvax/TTS/TransformSpeakerVoiceEvent.cs
@@ -1,0 +1,17 @@
+﻿using Content.Shared.Inventory;
+
+namespace Content.Shared.Corvax.TTS;
+
+public sealed class TransformSpeakerVoiceEvent : EntityEventArgs, IInventoryRelayEvent
+{
+    public SlotFlags TargetSlots => SlotFlags.MASK;
+    public EntityUid Sender;
+    public string VoiceId;
+
+    public TransformSpeakerVoiceEvent(EntityUid sender, string voiceId)
+    {
+        Sender = sender;
+        VoiceId = voiceId;
+    }
+
+}
